+++ src/devcenter/azext_devcenter/utils.py
@@ -5,7 +5,7 @@
 #
 # --------------------------------------------------------------------------
 import json
-from datetime import timedelta
+from datetime import (timedelta, datetime)
 from azure.cli.core.azclierror import ResourceNotFoundError, AzureInternalError
 from azure.cli.core.util import send_raw_request
 from azure.cli.core._profile import Profile
@@ -27,8 +27,10 @@ def get_project_arg(cli_ctx, dev_center_name, project_name=None):
     | take 1
     | extend devCenterUri = properties.devCenterUri
     | project name,devCenterUri """
+    options = {
+        "allowPartialScopes": True}  # maximum of 5000 subs for cross tenant query
 
-    content = {"query": query}
+    content = {"query": query, "options": options}
     request_url = f"{management_hostname}/providers/Microsoft.ResourceGraph/resources?api-version={api_version}"
 
     response = send_raw_request(
@@ -52,7 +54,8 @@ def get_project_data(cli_ctx, dev_center_name, project_name=None):
     profile = Profile()
     tenant_id = profile.get_subscription()["tenantId"]
 
-    resource_graph_data = get_project_arg(cli_ctx, dev_center_name, project_name)
+    resource_graph_data = get_project_arg(
+        cli_ctx, dev_center_name, project_name)
 
     error_help = f"""under the current tenant '{tenant_id}'. \
 Please contact your admin to gain access to specific projects or \
@@ -73,8 +76,9 @@ use a different tenant where you have access to projects."""
 def get_earliest_time(action_iterator):
     earliest_time = None
     for action in action_iterator:
-        if action.next is not None:
-            action_time = action.next.scheduled_time
+        if action["next"] is not None:
+            action_string = action["next"]["scheduledTime"]
+            action_time = datetime.strptime(action_string, "%Y-%m-%dT%H:%M:%S.%fZ")
             if earliest_time is None or action_time < earliest_time:
                 earliest_time = action_time
     return earliest_time