+++ src/devcenter/HISTORY.rst
@@ -2,6 +2,20 @@
 
 Release History
 ===============
+3.0.0
+++++++
+* Update control plane API to v2023-06-01-preview
+* Update data plane API to v2023-07-01-preview
+
+=======
+2.0.0
+++++++
+* Update dev-box-definition to devbox-definition parameter
+
+1.0.2
+++++++
+* Update allowPartialScopes to True for ARG query
+
 1.0.1
 ++++++
 * Update error string for environment create same name validation